QM Parser are removed from QMTP as methods
To improve encapsulation even more, all specific
and explicit parser methods in QMTP are removed.

Instead a generic parseOutput method is created that is
a wrapper method to specific parsers.

The wrapper method is run on two occasions:
1) if a successful QM result has found for a species, it calls it
2) if there is no previous QM result, first run the QM job,
 then call parser method

Determining whether a successsful QM result exists is now done by
checking if one of the flags fo QMverifier is set to true.