generalized runQM and createQMInput methods
instead of the specific methods in QMTP to create input files and
call QM programs, these methods have been removed and replaced by a
generic run and createQMInput methods

this greatly simplifies the generateQMThermoData method,which
was overloaded.

Note that due to the different syntax of the MM4HR input file writer,
this part is still there.