1,2-biradical perception and conversion into double bonds
After the creation of products in the ReactionAdjList class,
every chemgraph is checked whether it is a 1,2-biradical,
i.e. a structure with two adjacent monoradicals.

If a 1,2-biradical is encountered, the chemgraph is changed by:
- removing the 2 unpaired electrons (actually setting the order
of the FreeElectron to 0)
- increasing the bond order by 1.

Since this is simply a development commit, I contrained this approach
to:

- species with exactly two radical sites
- species with a single bond in between the two adjacent atoms

all of this is contained in a single private method.